Move vendored packages in `core` to `libs`
Remove `core` references from vendored packages in `core` and optimize imports
