+++ libs/linktastic/linktastic.py
@@ -22,9 +22,9 @@
 # WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
 # SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
+import os
 import subprocess
 from subprocess import CalledProcessError
-import os
 
 if os.name == 'nt':
     info = subprocess.STARTUPINFO()
