+++ libs/qbittorrent/client.py
@@ -1,5 +1,6 @@
-import requests
 import json
+
+import requests
 
 
 class LoginRequired(Exception):
