+++ libs/synchronousdeluge/client.py
@@ -1,9 +1,9 @@
 # coding=utf-8
 import os
 import platform
-
 from collections import defaultdict
 from itertools import imap
+
 from .exceptions import DelugeRPCError
 from .protocol import DelugeRPCRequest, DelugeRPCResponse
 from .transfer import DelugeTransfer
