+++ libs/synchronousdeluge/exceptions.py
