+++ libs/transmissionrpc/constants.py
